Move module to its own place
