+++ apps/transport/lib/config/runtime_configuration.ex
@@ -1,0 +1,24 @@
+defmodule Transport.RuntimeConfiguration do
+  def build_config(env_provider, config_env) do
+    case config_env do
+      :prod ->
+        {
+          env_provider.get_env("WORKER") || raise("expected the WORKER environment variable to be set"),
+          env_provider.get_env("WEBSERVER") || raise("expected the WEBSERVER variable to be set")
+        }
+
+      :dev ->
+        # By default in dev, the application will be both a worker and a webserver
+        {
+          env_provider.get_env("WORKER", "1"),
+          env_provider.get_env("WEBSERVER", "1")
+        }
+
+      :test ->
+        {
+          "0",
+          "0"
+        }
+    end
+  end
+end
